Replace SWR by tPRC
