+++ package.json
@@ -39,7 +39,6 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "superjson": "^1.11.0",
-    "swr": "^1.3.0",
     "zod": "^3.19.1"
   },
   "devDependencies": {
