+++ src/components/DashboardIncomes.tsx
@@ -1,16 +1,13 @@
 import { Loader, SimpleGrid } from "@mantine/core";
-import useSWR from "swr";
-import { ApiGetDashboardIncomes } from "../server/dashboard";
+import { trpc } from "../lib/trpc";
 import AlertFetchError from "./AlertFetchError";
 import StatsIncomeLastThreeMonths from "./StatsIncomeLastThreeMonths";
 
 export default function DashboardIncomes() {
-  const { data, error } = useSWR<ApiGetDashboardIncomes>(
-    "/api/dashboard/incomes"
-  );
+  const { data, error, isLoading } = trpc.dashboard.compareIncomes.useQuery();
 
   if (error) return <AlertFetchError />;
-  if (!data) return <Loader />;
+  if (isLoading) return <Loader />;
 
   return (
     <SimpleGrid cols={data.length}>
