+++ src/pages/_app.tsx
@@ -7,7 +7,6 @@ import { Session } from "next-auth";
 import { SessionProvider } from "next-auth/react";
 import { AppProps } from "next/app";
 import Head from "next/head";
-import { SWRConfig } from "swr";
 import { CategoriesProvider } from "../contexts/CategoriesContext";
 import { defaultProps } from "../lib/mantine";
 import { trpc } from "../lib/trpc";
@@ -53,25 +52,11 @@ export default trpc.withTRPC(function App(
       >
         <ModalsProvider>
           <CategoriesProvider>
-            <SWRConfig
-              value={{
-                fetcher: async (url) => {
-                  const res = await fetch(url);
-                  if (!res.ok) {
-                    throw new Error(
-                      "An error occurred while fetching the data."
-                    );
-                  }
-                  return res.json();
-                },
-              }}
-            >
-              <NotificationsProvider>
-                <SessionProvider session={pageProps.session}>
-                  <Component {...pageProps} />
-                </SessionProvider>
-              </NotificationsProvider>
-            </SWRConfig>
+            <NotificationsProvider>
+              <SessionProvider session={pageProps.session}>
+                <Component {...pageProps} />
+              </SessionProvider>
+            </NotificationsProvider>
           </CategoriesProvider>
         </ModalsProvider>
       </MantineProvider>
