+++ src/server/trpc/_app.ts
@@ -1,9 +1,11 @@
 import { router } from ".";
 import { categoriesRouter } from "./categories";
+import { dashboardRouter } from "./dashboard";
 import { transactionRouter } from "./transactions";
 
 export const appRouter = router({
   categories: categoriesRouter,
+  dashboard: dashboardRouter,
   transactions: transactionRouter,
 });
