+++ src/utils/constant.ts
@@ -36,6 +36,8 @@ export const MODULES = {
   },
 };
 
+export const MONTH_YEAR = "MMM YYYY";
+
 // TODO: Add currency and locale in user settings
 export const CURRENCY = "CAD";
 export const LOCALE = "en-US";
