+++ src/utils/getIncomeCategories.ts
@@ -1,0 +1,8 @@
+import { CategoryType, PrismaClient } from "@prisma/client";
+
+export async function getIncomeCategories(prisma: PrismaClient) {
+  return prisma.category.findMany({
+    select: { id: true, name: true },
+    where: { name: { contains: "Deposit" }, type: CategoryType.Income },
+  });
+}
